+++ src/Controller/Admin/Operator/OperatorWorkRegisterHeaderAdminController.php
@@ -85,10 +85,10 @@ class OperatorWorkRegisterHeaderAdminController extends BaseAdminController
                 if (str_contains($singleResult['description'], 'Hora laboral')) {
                     $result['workingHour'] += $singleResult['hours'];
                 }
-                if (str_contains($singleResult['description'], 'Hora normal')) {
+                if (str_contains($singleResult['description'], 'Hora extra')) {
                     $result['normalHour'] += $singleResult['hours'];
                 }
-                if (str_contains($singleResult['description'], 'Hora extra')) {
+                if (str_contains($singleResult['description'], 'Hora nocturna')) {
                     $result['extraHour'] += $singleResult['hours'];
                 }
                 if (str_contains($singleResult['description'], 'Hora negativa')) {