+++ src/Controller/Admin/Sale/SaleInvoiceAdminController.php
@@ -46,7 +46,7 @@ class SaleInvoiceAdminController extends BaseAdminController
     {
         $saleInvoices = $selectedModelQuery->execute()->getQuery()->getResult();
         usort($saleInvoices, function (SaleInvoice $a, SaleInvoice $b) {
-            return $a->getDateFormatted() > $b->getDateFormatted();
+            return $a->getDate() <=> $b->getDate();
         });
         $siforDates = $saleInvoices;
         $filterInfo = $this->admin->getFilterParameters();
@@ -71,7 +71,7 @@ class SaleInvoiceAdminController extends BaseAdminController
     {
         $saleInvoices = $selectedModelQuery->execute()->getQuery()->getResult();
         usort($saleInvoices, function (SaleInvoice $a, SaleInvoice $b) {
-            return $a->getDateFormatted() > $b->getDateFormatted();
+            return $a->getDate() <=> $b->getDate();
         });
         $siforDates = $saleInvoices;
         $filterInfo = $this->admin->getFilterParameters();
@@ -266,6 +266,22 @@ class SaleInvoiceAdminController extends BaseAdminController
         return parent::batchActionDelete($query);
     }
 
+    public function batchActionHasBeenCounted(ProxyQueryInterface $query): Response
+    {
+        $saleInvoices = $query->execute();
+        try {
+            /** @var SaleInvoice $saleInvoice */
+            foreach ($saleInvoices as $saleInvoice) {
+                $saleInvoice->setHasBeenCounted(true);
+                    $this->admin->getModelManager()->update($saleInvoice);
+            }
+        } catch (\Throwable $ex) {
+            $this->addFlash('warning', 'No se pudo realizar la acción. Error: ' . $ex->getMessage().' Trace: ' . $ex->getTraceAsString());
+        }
+
+        return $this->redirectToRoute('admin_app_sale_saleinvoice_list');
+    }
+
     /**
      * @throws ModelManagerException
      * @throws \Sonata\AdminBundle\Exception\ModelManagerThrowable
@@ -344,11 +360,7 @@ class SaleInvoiceAdminController extends BaseAdminController
         if (!$saleInvoice) {
             throw $this->createNotFoundException(sprintf('unable to find the object with id: %s', $id));
         }
-
-        $xml = $this->EFacturaService->createEFactura(
-            $saleInvoice,
-            billingPeriodStart: $saleInvoice->getDate(),
-            billingPeriodEnd: $saleInvoice->getDate());
+        $xml = $this->im->createEInvoice($saleInvoice);
         $response = new Response($xml);
         $response->headers->set('Content-type', 'text/xml');
         $response->headers->set('Content-Disposition', 'attachment; filename="factura-e-'.$saleInvoice->getInvoiceNumber().'.xml"');